requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
curl_cffi==0.11.1
discord-protos==0.0.2
frozenlist==1.6.0
idna==3.10
multidict==6.4.4
propcache==0.3.1
protobuf==6.31.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
selfcord.py @ git+https://github.com/dolfies/discord.py-self@c9ae978b02661b385c46de5936ce1618cc6b113e
tzdata==2025.2
tzlocal==5.3.1
yarl==1.20.0
